test: refactored utility actions
