+++ src/actions/index.ts
@@ -1,12 +1,9 @@
 export * from './transactions';
 export * from './initStore';
 export * from './initStoreV2';
-export * from './makeAuction';
 export * from './mintNFT';
 export * from './mintEditionFromMaster';
 export * from './closeVault';
-export * from './createVault';
-export * from './createExternalPriceAccount';
 export * from './createMetadata';
 export * from './createMasterEdition';
 export * from './signMetadata';
