+++ src/actions/utility/createVault.ts
@@ -1,13 +1,13 @@
-import { Connection } from '../Connection';
-import { Wallet } from '../wallet';
-
 import { Transaction } from '@metaplex-foundation/mpl-core';
+import { AccountLayout, MintLayout, NATIVE_MINT } from '@solana/spl-token';
 import { InitVault, Vault, VaultProgram } from '@metaplex-foundation/mpl-token-vault';
 import { Keypair, PublicKey, SystemProgram, TransactionSignature } from '@solana/web3.js';
-import { AccountLayout, MintLayout, NATIVE_MINT } from '@solana/spl-token';
-import { CreateMint, CreateTokenAccount } from '../programs';
-import { sendTransaction } from '../actions/transactions';
-import { TransactionsBatch } from '../utils/transactions-batch';
+
+import { Wallet } from '../../wallet';
+import { Connection } from '../../Connection';
+import { sendTransaction } from '../transactions';
+import { CreateMint, CreateTokenAccount } from '../../programs';
+import { TransactionsBatch } from '../../utils/transactions-batch';
 
 interface CreateVaultParams {
   connection: Connection;
